requirements.txt
mistralai>=1.0.0
anthropic>=0.39.0
faiss-cpu>=1.8.0
numpy>=2.0.0
pandas>=2.2.0
rapidfuzz>=3.9.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
matplotlib>=3.9.0
pytest>=8.0.0
pytest-cov>=5.0.0
httpx>=0.27.0
